Add tool for formatting short courses taught
This is a relevant section on the promotion CV so I may as well add it
to my resume.  I now have a nice YAML format that should be easy to
maintain and a script to massage it into a proper table format.  The
primary output is LaTeX because that is the CV format.  At the moment,
I'm relying on Pandoc to format the HTML table.  I may change this in
the future and add Markdown as a proper output of the script.

I also had to conditionally define the Environment to mainly check if we
need to define the default target.